Add comments

+++ Publishing-Service-Management-Information-System/src/main/java/com/aivarasnakvosas/publishingservicemis/config/properties/SecurityProperties.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Component;
 public class SecurityProperties {
 
     private String jwtTokenSecret;
-    private Integer jwtExpirationTime;
+    private Long jwtExpirationTime;
 }
